move msw to js to remove ts-node from our process
