+++ mocks/index.js
@@ -1,3 +1,9 @@
-require("tsconfig-paths/register");
-require("ts-node").register({ transpileOnly: true });
-require("./start");
+import { setupServer } from "msw/node";
+
+const server = setupServer();
+
+server.listen({ onUnhandledRequest: "bypass" });
+console.info("🔶 Mock server running");
+
+process.once("SIGINT", () => server.close());
+process.once("SIGTERM", () => server.close());
