--- mocks/start.ts
@@ -1,9 +1,0 @@
-import { setupServer } from "msw/node";
-
-const server = setupServer();
-
-server.listen({ onUnhandledRequest: "bypass" });
-console.info("🔶 Mock server running");
-
-process.once("SIGINT", () => server.close());
-process.once("SIGTERM", () => server.close());
